dev(arch): refactor x86
