+++ src/arch/x86/idt.c
@@ -1,11 +1,4 @@
-#include <libft.h>
-
-#include "idt.h"
-#include "io.h"
-#include "keyboard.h"
-#include "panic.h"
-#include "vga.h"
-#include <acpi.h>
+#include "internal/idt.h"
 
 #define PIC1         0x20 // IO base address for master PIC
 #define PIC2         0xA0 // IO base address for slave PIC
@@ -151,6 +144,8 @@ static inline void idt_set_entry(idt_entry *ptr, uint16_t selector, uint8_t type
 	ptr->selector        = selector;
 	ptr->type_attributes = type;
 }
+
+const idtr_t *idtr_getter(void) { return &idtr; }
 
 void idt_init(void)
 {
