+++ src/arch/x86/internal/x86.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void gdt_init(void);
+void idt_init(void);
