+++ src/arch/x86/x86.c
@@ -1,0 +1,7 @@
+#include "internal/x86.h"
+
+void x86_init(void)
+{
+	gdt_init();
+	idt_init();
+}
