+++ src/kernel/internal/panic.h
@@ -4,9 +4,9 @@
 // IMCLUDES
 // ============================================================================
 
-#include "panic.h"
 #include <acpi.h>
 #include <drivers/vga.h>
+#include <kernel/panic.h>
 #include <types.h>
 
 // ============================================================================
