+++ src/memory/buddy.c
@@ -1,9 +1,5 @@
 #include "internal/buddy.h"
-#include "internal/boot_allocator.h"
-#include "panic.h"
-#include "utils.h"
-#include "vga.h"
-#include <libft.h>
+// #include "internal/boot_allocator.h"
 
 /////////////////////////////////////////////////////////////////////////////////
 // Internal APIs
