+++ src/memory/internal/buddy.h
@@ -5,7 +5,12 @@
 // ============================================================================
 
 #include <acpi.h>
+#include <drivers/vga.h>
+#include <kernel/panic.h>
+#include <libft.h>
+#include <memory/boot_allocator.h>
 #include <memory/buddy.h>
+#include <utils.h>
 
 // ============================================================================
 // DEFINE AND MACRO
